Refactored ImageManip
