+++ include/depthai/utility/ImageManipImpl.hpp
@@ -48,14 +48,14 @@
 
 namespace dai {
 namespace impl {
-template <typename N, template <typename T> typename ImageManipBuffer>
+template <typename N>
 void loop(N& node,
           const ImageManipConfig& initialConfig,
           std::shared_ptr<spdlog::async_logger> logger,
-          std::function<size_t(const ImageManipConfig&, const ImgFrame&)> build,
-          std::function<std::shared_ptr<ImgFrame>(size_t)> getFrame,
-          std::function<bool(const std::shared_ptr<OffsetMemory>&, std::shared_ptr<OffsetMemory>)> apply,
-          std::function<void(const ImgFrame&, ImgFrame&)> setFrame) {
+          const std::function<size_t(const ImageManipConfig&, const ImgFrame&)>& build,
+          const std::function<std::shared_ptr<ImgFrame>(size_t)>& getFrame,
+          const std::function<bool(const std::shared_ptr<OffsetMemory>&, std::shared_ptr<OffsetMemory>)>& apply,
+          const std::function<void(const ImgFrame&, ImgFrame&)>& setFrame) {
     using namespace std::chrono;
     auto config = initialConfig;
 
@@ -125,7 +125,7 @@ void loop(N& node,
             bool success = true;
             {
                 auto t3 = steady_clock::now();
-                success = apply(inImage->data, outImage->data);
+                success = apply(ConvertedOffsetMemory::convert(inImage->data), ConvertedOffsetMemory::convert(outImage->data));
                 auto t4 = steady_clock::now();
 
                 setFrame(*inImage, *outImage);
@@ -402,75 +402,51 @@ class WarpH : public Warp {
     void apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst) override;
 };
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
 class ColorChange {
+   protected:
     std::shared_ptr<spdlog::async_logger> logger;
 
-    std::shared_ptr<ImageManipData> ccAuxFrame;
     ImgFrame::Type from;
     ImgFrame::Type to;
 
     FrameSpecs srcSpecs;
     FrameSpecs dstSpecs;
 
-    bool colorConvertToRGB888i(const std::shared_ptr<ImageManipData> inputFrame,
-                               std::shared_ptr<ImageManipData> outputFrame,
-                               FrameSpecs srcSpecs,
-                               FrameSpecs dstSpecs,
-                               ImgFrame::Type from);
-    bool colorConvertToBGR888p(const std::shared_ptr<ImageManipData> inputFrame,
-                               std::shared_ptr<ImageManipData> outputFrame,
-                               FrameSpecs srcSpecs,
-                               FrameSpecs dstSpecs,
-                               ImgFrame::Type from);
-    bool colorConvertToRGB888p(const std::shared_ptr<ImageManipData> inputFrame,
-                               std::shared_ptr<ImageManipData> outputFrame,
-                               FrameSpecs srcSpecs,
-                               FrameSpecs dstSpecs,
-                               ImgFrame::Type from);
-    bool colorConvertToBGR888i(const std::shared_ptr<ImageManipData> inputFrame,
-                               std::shared_ptr<ImageManipData> outputFrame,
-                               FrameSpecs srcSpecs,
-                               FrameSpecs dstSpecs,
-                               ImgFrame::Type from);
-    bool colorConvertToNV12(const std::shared_ptr<ImageManipData> inputFrame,
-                            std::shared_ptr<ImageManipData> outputFrame,
-                            FrameSpecs srcSpecs,
-                            FrameSpecs dstSpecs,
-                            ImgFrame::Type from);
-    bool colorConvertToYUV420p(const std::shared_ptr<ImageManipData> inputFrame,
-                               std::shared_ptr<ImageManipData> outputFrame,
-                               FrameSpecs srcSpecs,
-                               FrameSpecs dstSpecs,
-                               ImgFrame::Type from);
-    bool colorConvertToGRAY8(const std::shared_ptr<ImageManipData> inputFrame,
-                             std::shared_ptr<ImageManipData> outputFrame,
-                             FrameSpecs srcSpecs,
-                             FrameSpecs dstSpecs,
-                             ImgFrame::Type from);
-    bool colorConvert(const std::shared_ptr<ImageManipData> inputFrame,
-                      std::shared_ptr<ImageManipData> outputFrame,
+   public:
+    ColorChange() = default;
+    ColorChange(std::shared_ptr<spdlog::async_logger> logger) : logger(logger) {}
+    virtual ~ColorChange() = default;
+
+    void setLogger(std::shared_ptr<spdlog::async_logger> logger) {
+        this->logger = logger;
+    }
+
+    virtual void build(const FrameSpecs srcFrameSpecs, const FrameSpecs dstFrameSpecs, const ImgFrame::Type typeFrom, const ImgFrame::Type typeTo) = 0;
+
+    virtual void apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst) = 0;
+};
+
+class ColorChangeH : public ColorChange {
+    std::shared_ptr<_ImageManipMemory> ccAuxFrame;
+
+    bool colorConvert(const std::shared_ptr<_ImageManipMemory> inputFrame,
+                      std::shared_ptr<_ImageManipMemory> outputFrame,
                       FrameSpecs srcSpecs,
                       FrameSpecs dstSpecs,
                       ImgFrame::Type from,
                       ImgFrame::Type to);
 
    public:
-    ColorChange() = default;
-    ColorChange(std::shared_ptr<spdlog::async_logger> logger) : logger(logger) {}
-
-    void setLogger(std::shared_ptr<spdlog::async_logger> logger) {
-        this->logger = logger;
-    }
-
     void build(const FrameSpecs srcFrameSpecs, const FrameSpecs dstFrameSpecs, const ImgFrame::Type typeFrom, const ImgFrame::Type typeTo);
 
     void apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst);
 };
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
 class ImageManipOperations {
+    static_assert(std::is_base_of<OffsetMemory, ImageManipData>::value, "ImageManipData must be derived from OffsetMemory");
     static_assert(std::is_base_of<Warp, WarpBackend>::value, "WarpBackend must be derived from Warp");
+    static_assert(std::is_base_of<ColorChange, ColorChangeBackend>::value, "ColorChangeBackend must be derived from ColorChange");
     using Container = std::vector<ManipOp>;
 
     static constexpr uint8_t MODE_CONVERT = 1;
@@ -503,9 +479,9 @@ class ImageManipOperations {
 
     FrameSpecs srcSpecs;
 
-    ColorChange<ImageManipBuffer, ImageManipData> preprocCc;
+    ColorChangeBackend preprocCc;
     WarpBackend warpEngine;
-    ColorChange<ImageManipBuffer, ImageManipData> clrChange;
+    ColorChangeBackend clrChange;
 
    public:
     ImageManipOperations(ImageManipProperties props, std::shared_ptr<spdlog::async_logger> logger = nullptr) : properties(props), logger(logger) {
@@ -546,8 +522,50 @@ class ImageManipOperations {
     std::string toString() const;
 };
 
+struct ColorChangeArgs {
+    const std::shared_ptr<OffsetMemory>& inputFrame;
+    std::shared_ptr<OffsetMemory> outputFrame;
+    FrameSpecs srcSpecs;
+    FrameSpecs dstSpecs;
+    ImgFrame::Type from;
+    std::shared_ptr<OffsetMemory> auxFrame;
+};
 FrameSpecs getSrcFrameSpecs(dai::ImgFrame::Specs srcSpecs);
+FrameSpecs getCcDstFrameSpecs(FrameSpecs srcSpecs, dai::ImgFrame::Type from, dai::ImgFrame::Type to);
+FrameSpecs getDstFrameSpecs(size_t width, size_t height, dai::ImgFrame::Type type);
 size_t getAlignedOutputFrameSize(ImgFrame::Type type, size_t width, size_t height);
+void printSpecs(spdlog::async_logger& logger, FrameSpecs specs);
+size_t getFrameSize(const ImgFrame::Type type, const FrameSpecs& specs);
+
+bool isTypeSupported(dai::ImgFrame::Type type);
+
+bool colorConvertToRGB888p(const ColorChangeArgs& args);
+bool colorConvertToBGR888p(const ColorChangeArgs& args);
+bool colorConvertToRGB888i(const ColorChangeArgs& args);
+bool colorConvertToBGR888i(const ColorChangeArgs& args);
+bool colorConvertToNV12(const ColorChangeArgs& args);
+bool colorConvertToYUV420p(const ColorChangeArgs& args);
+bool colorConvertToGRAY8(const ColorChangeArgs& args);
+
+std::tuple<float, float, float, float> getOuterRect(const std::vector<std::array<float, 2>> points);
+dai::RotatedRect getOuterRotatedRect(const std::vector<std::array<float, 2>>& points);
+std::array<std::array<float, 3>, 3> getResizeMat(Resize o, float width, float height, uint32_t outputWidth, uint32_t outputHeight);
+void getOutputSizeFromCorners(const std::array<std::array<float, 2>, 4>& corners,
+                              const bool center,
+                              const std::array<std::array<float, 3>, 3> transformInv,
+                              const uint32_t srcWidth,
+                              const uint32_t srcHeight,
+                              uint32_t& outputWidth,
+                              uint32_t& outputHeight);
+
+void getTransformImpl(const ManipOp& op,
+                      std::array<std::array<float, 3>, 3>& transform,
+                      std::array<std::array<float, 2>, 4>& imageCorners,
+                      std::vector<std::array<std::array<float, 2>, 4>>& srcCorners,
+                      uint32_t& outputWidth,
+                      uint32_t& outputHeight);
+
+
 
 }  // namespace impl
 }  // namespace dai
@@ -565,1967 +583,11 @@ constexpr T ALIGN_UP(T value, std::size_t alignment) {
 }
 #endif
 
-FrameSpecs getSrcFrameSpecs(dai::ImgFrame::Specs srcSpecs);
-FrameSpecs getCcDstFrameSpecs(FrameSpecs srcSpecs, dai::ImgFrame::Type from, dai::ImgFrame::Type to);
-FrameSpecs getDstFrameSpecs(size_t width, size_t height, dai::ImgFrame::Type type);
-
-void transformOpenCV(const uint8_t* src,
-                     uint8_t* dst,
-                     const size_t srcWidth,
-                     const size_t srcHeight,
-                     const size_t srcStride,
-                     const size_t dstWidth,
-                     const size_t dstHeight,
-                     const size_t dstStride,
-                     const uint16_t numChannels,
-                     const uint16_t bpp,
-                     const std::array<std::array<float, 3>, 3> matrix,
-                     const std::vector<uint32_t>& background,
-                     const FrameSpecs& srcImgSpecs,
-                     const size_t sourceMinX,
-                     const size_t sourceMinY,
-                     const size_t sourceMaxX,
-                     const size_t sourceMaxY);
-void transformFastCV(const uint8_t* src,
-                     uint8_t* dst,
-                     const size_t srcWidth,
-                     const size_t srcHeight,
-                     const size_t srcStride,
-                     const size_t dstWidth,
-                     const size_t dstHeight,
-                     const size_t dstStride,
-                     const uint16_t numChannels,
-                     const uint16_t bpp,
-                     const std::array<std::array<float, 3>, 3> matrix,
-                     const std::vector<uint32_t>& background,
-                     const FrameSpecs& srcImgSpecs,
-                     const size_t sourceMinX,
-                     const size_t sourceMinY,
-                     const size_t sourceMaxX,
-                     const size_t sourceMaxY,
-                     uint32_t* fastCvBorder);
-
-static inline int clampi(int val, int minv, int maxv) {
-    // return val < minv ? minv : (val > maxv ? maxv : val);
-    return std::clamp(val, minv, maxv);
-}
-static inline float clampf(float val, float minv, float maxv) {
-    // return val < minv ? minv : (val > maxv ? maxv : val);
-    return std::clamp(val, minv, maxv);
-}
-
-bool isTypeSupported(dai::ImgFrame::Type type);
-
-bool getFrameTypeInfo(dai::ImgFrame::Type outFrameType, int& outNumPlanes, float& outBpp);
-
-//-----------------------------
-//--- Frame type conversion ---
-//-----------------------------
-
-static inline void YUVfromRGB(float& Y, float& U, float& V, const float R, const float G, const float B) {
-    Y = 0.257f * R + 0.504f * G + 0.098f * B + 16;
-    U = -0.148f * R - 0.291f * G + 0.439f * B + 128;
-    V = 0.439f * R - 0.368f * G - 0.071f * B + 128;
-}
-static inline void RGBfromYUV(float& R, float& G, float& B, float Y, float U, float V) {
-    Y -= 16;
-    U -= 128;
-    V -= 128;
-    R = 1.164f * Y + 1.596f * V;
-    G = 1.164f * Y - 0.392f * U - 0.813f * V;
-    B = 1.164f * Y + 2.017f * U;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToRGB888p(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                          std::shared_ptr<ImageManipData> outputFrame,
-                                                                          FrameSpecs srcSpecs,
-                                                                          FrameSpecs dstSpecs,
-                                                                          dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::RGB888p;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case dai::ImgFrame::Type::BGR888p:
-            std::copy(src + srcSpecs.p1Offset, src + srcSpecs.p2Offset, outputFrame->data() + dstSpecs.p3Offset);
-            std::copy(src + srcSpecs.p2Offset, src + srcSpecs.p3Offset, outputFrame->data() + dstSpecs.p2Offset);
-            std::copy(src + srcSpecs.p3Offset, src + inputSize, outputFrame->data() + dstSpecs.p1Offset);
-            done = true;
-            break;
-        case dai::ImgFrame::Type::RGB888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat img(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            cv::split(img, channels);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t srcPos = lineStart + j * 3;
-                    uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    uint32_t p2Pos = dstSpecs.p2Offset + i * dstSpecs.p2Stride + j;
-                    uint32_t p3Pos = dstSpecs.p3Offset + i * dstSpecs.p3Stride + j;
-                    outputFrame->getOffsetData()[p1Pos] = src[srcPos + 0];
-                    outputFrame->getOffsetData()[p2Pos] = src[srcPos + 1];
-                    outputFrame->getOffsetData()[p3Pos] = src[srcPos + 2];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat img(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::split(img, channels);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t srcPos = lineStart + j * 3;
-                    uint32_t p1Pos = dstSpecs.p3Offset + i * dstSpecs.p3Stride + j;
-                    uint32_t p2Pos = dstSpecs.p2Offset + i * dstSpecs.p2Stride + j;
-                    uint32_t p3Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    outputFrame->getOffsetData()[p1Pos] = src[srcPos + 0];
-                    outputFrame->getOffsetData()[p2Pos] = src[srcPos + 1];
-                    outputFrame->getOffsetData()[p3Pos] = src[srcPos + 2];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::NV12: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PseudoPlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                                   src + srcSpecs.p2Offset,
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   srcSpecs.p1Stride,
-                                                   srcSpecs.p2Stride,
-                                                   ccAuxFrame->data(),
-                                                   auxStride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat frameUV(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC2, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            cv::Mat auxBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::cvtColorTwoPlane(frameY, frameUV, auxBGR, cv::COLOR_YUV2BGR_NV12);
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::split(auxBGR, channels);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::YUV420p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                             src + srcSpecs.p2Offset,
-                                             src + srcSpecs.p3Offset,
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             srcSpecs.p1Stride,
-                                             srcSpecs.p2Stride,
-                                             srcSpecs.p3Stride,
-                                             ccAuxFrame->data(),
-                                             auxStride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartY = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartU = srcSpecs.p2Offset + (i / 2) * srcSpecs.p2Stride;
-                const uint32_t lineStartV = srcSpecs.p3Offset + (i / 2) * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + i * dstSpecs.p2Stride + j;
-                    const uint32_t p3Pos = dstSpecs.p3Offset + i * dstSpecs.p3Stride + j;
-                    float Y = src[lineStartY + j];
-                    float U = src[lineStartU + (uint32_t)(j / 2)];
-                    float V = src[lineStartV + (uint32_t)(j / 2)];
-                    float R, G, B;
-                    RGBfromYUV(R, G, B, Y, U, V);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(clampi(roundf(R), 0, 255));
-                    outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(clampi(roundf(G), 0, 255));
-                    outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(clampi(roundf(B), 0, 255));
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAY8:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToBGR888p(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                          std::shared_ptr<ImageManipData> outputFrame,
-                                                                          FrameSpecs srcSpecs,
-                                                                          FrameSpecs dstSpecs,
-                                                                          dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::BGR888p;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p:
-            std::copy(src + srcSpecs.p1Offset, src + srcSpecs.p2Offset, outputFrame->data() + dstSpecs.p3Offset);
-            std::copy(src + srcSpecs.p2Offset, src + srcSpecs.p3Offset, outputFrame->data() + dstSpecs.p2Offset);
-            std::copy(src + srcSpecs.p3Offset, src + inputSize, outputFrame->data() + dstSpecs.p1Offset);
-            done = true;
-            break;
-        case dai::ImgFrame::Type::BGR888p:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case dai::ImgFrame::Type::RGB888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat img(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::split(img, channels);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t srcPos = lineStart + j * 3;
-                    uint32_t p1Pos = dstSpecs.p3Offset + i * dstSpecs.p3Stride + j;
-                    uint32_t p2Pos = dstSpecs.p2Offset + i * dstSpecs.p2Stride + j;
-                    uint32_t p3Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    outputFrame->getOffsetData()[p1Pos] = src[srcPos + 0];
-                    outputFrame->getOffsetData()[p2Pos] = src[srcPos + 1];
-                    outputFrame->getOffsetData()[p3Pos] = src[srcPos + 2];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat img(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            cv::split(img, channels);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t srcPos = lineStart + j * 3;
-                    uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    uint32_t p2Pos = dstSpecs.p2Offset + i * dstSpecs.p2Stride + j;
-                    uint32_t p3Pos = dstSpecs.p3Offset + i * dstSpecs.p3Stride + j;
-                    outputFrame->getOffsetData()[p1Pos] = src[srcPos + 0];
-                    outputFrame->getOffsetData()[p2Pos] = src[srcPos + 1];
-                    outputFrame->getOffsetData()[p3Pos] = src[srcPos + 2];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::NV12: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PseudoPlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                                   src + srcSpecs.p2Offset,
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   srcSpecs.p1Stride,
-                                                   srcSpecs.p2Stride,
-                                                   ccAuxFrame->data(),
-                                                   auxStride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat frameUV(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC2, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            cv::Mat auxBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::cvtColorTwoPlane(frameY, frameUV, auxBGR, cv::COLOR_YUV2BGR_NV12);
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            cv::split(auxBGR, channels);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::YUV420p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                             src + srcSpecs.p2Offset,
-                                             src + srcSpecs.p3Offset,
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             srcSpecs.p1Stride,
-                                             srcSpecs.p2Stride,
-                                             srcSpecs.p3Stride,
-                                             ccAuxFrame->data(),
-                                             auxStride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_2,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_1,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(ccAuxFrame->data(),
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                auxStride,
-                                0,
-                                0,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_RGB,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartY = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartU = srcSpecs.p2Offset + (i / 2) * srcSpecs.p2Stride;
-                const uint32_t lineStartV = srcSpecs.p3Offset + (i / 2) * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + i * dstSpecs.p2Stride + j;
-                    const uint32_t p3Pos = dstSpecs.p3Offset + i * dstSpecs.p3Stride + j;
-                    float Y = src[lineStartY + j];
-                    float U = src[lineStartU + (uint32_t)(j / 2)];
-                    float V = src[lineStartV + (uint32_t)(j / 2)];
-                    float R, G, B;
-                    RGBfromYUV(R, G, B, Y, U, V);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(clampi(roundf(B), 0, 255));
-                    outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(clampi(roundf(G), 0, 255));
-                    outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(clampi(roundf(R), 0, 255));
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAY8:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToRGB888i(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                          std::shared_ptr<ImageManipData> outputFrame,
-                                                                          FrameSpecs srcSpecs,
-                                                                          FrameSpecs dstSpecs,
-                                                                          dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::RGB888i;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p1Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p1Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p3Offset,
-                                       srcSpecs.p3Stride,
-                                       outputFrame->data() + dstSpecs.p1Offset,
-                                       dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            cv::Mat img(dstSpecs.height, dstSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::merge(channels, img);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = dstSpecs.p1Offset + i * dstSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t dstPos = lineStart + j * 3;
-                    uint32_t p1Pos = srcSpecs.p1Offset + i * srcSpecs.p1Stride + j;
-                    uint32_t p2Pos = srcSpecs.p2Offset + i * srcSpecs.p2Stride + j;
-                    uint32_t p3Pos = srcSpecs.p3Offset + i * srcSpecs.p3Stride + j;
-                    outputFrame->getOffsetData()[dstPos + 0] = src[p1Pos];
-                    outputFrame->getOffsetData()[dstPos + 1] = src[p2Pos];
-                    outputFrame->getOffsetData()[dstPos + 2] = src[p3Pos];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p3Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p3Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p1Offset,
-                                       srcSpecs.p1Stride,
-                                       outputFrame->data() + dstSpecs.p1Offset,
-                                       dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat img(dstSpecs.height, dstSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::merge(channels, img);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = dstSpecs.p1Offset + i * dstSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t dstPos = lineStart + j * 3;
-                    uint32_t p1Pos = srcSpecs.p3Offset + i * srcSpecs.p3Stride + j;
-                    uint32_t p2Pos = srcSpecs.p2Offset + i * srcSpecs.p2Stride + j;
-                    uint32_t p3Pos = srcSpecs.p1Offset + i * srcSpecs.p1Stride + j;
-                    outputFrame->getOffsetData()[dstPos + 0] = src[p1Pos];
-                    outputFrame->getOffsetData()[dstPos + 1] = src[p2Pos];
-                    outputFrame->getOffsetData()[dstPos + 2] = src[p3Pos];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::RGB888i:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case dai::ImgFrame::Type::BGR888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToBGR888u8(
-                src + srcSpecs.p1Offset, srcSpecs.width, srcSpecs.height, srcSpecs.p1Stride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat img(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat imgBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(img, imgBGR, cv::COLOR_RGB2BGR);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStartSrc = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                uint32_t lineStartDst = dstSpecs.p1Offset + i * dstSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t dstPos = lineStartDst + j * 3;
-                    uint32_t srcPos = lineStartSrc + j * 3;
-                    outputFrame->getOffsetData()[dstPos + 0] = src[srcPos + 2];
-                    outputFrame->getOffsetData()[dstPos + 1] = src[srcPos + 1];
-                    outputFrame->getOffsetData()[dstPos + 2] = src[srcPos + 0];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::NV12: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PseudoPlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                                   src + srcSpecs.p2Offset,
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   srcSpecs.p1Stride,
-                                                   srcSpecs.p2Stride,
-                                                   outputFrame->data() + dstSpecs.p1Offset,
-                                                   dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat frameUV(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC2, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            cv::Mat auxBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::cvtColorTwoPlane(frameY, frameUV, auxBGR, cv::COLOR_YUV2BGR_NV12);
-            cv::Mat img(dstSpecs.height, dstSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(auxBGR, img, cv::COLOR_RGB2BGR);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::YUV420p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                             src + srcSpecs.p2Offset,
-                                             src + srcSpecs.p3Offset,
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             srcSpecs.p1Stride,
-                                             srcSpecs.p2Stride,
-                                             srcSpecs.p3Stride,
-                                             outputFrame->data() + dstSpecs.p1Offset,
-                                             dstSpecs.p1Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartY = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartU = srcSpecs.p2Offset + (i / 2) * srcSpecs.p2Stride;
-                const uint32_t lineStartV = srcSpecs.p3Offset + (i / 2) * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + 3 * j;
-                    float Y = src[lineStartY + j];
-                    float U = src[lineStartU + (uint32_t)(j / 2)];
-                    float V = src[lineStartV + (uint32_t)(j / 2)];
-                    float R, G, B;
-                    RGBfromYUV(R, G, B, Y, U, V);
-                    outputFrame->getOffsetData()[pos + 0] = static_cast<uint8_t>(clampi(roundf(R), 0, 255.0f));
-                    outputFrame->getOffsetData()[pos + 1] = static_cast<uint8_t>(clampi(roundf(G), 0, 255.0f));
-                    outputFrame->getOffsetData()[pos + 2] = static_cast<uint8_t>(clampi(roundf(B), 0, 255.0f));
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAY8:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToBGR888i(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                          std::shared_ptr<ImageManipData> outputFrame,
-                                                                          FrameSpecs srcSpecs,
-                                                                          FrameSpecs dstSpecs,
-                                                                          dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::BGR888i;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-#endif
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p3Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p3Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p1Offset,
-                                       srcSpecs.p1Stride,
-                                       outputFrame->data() + dstSpecs.p1Offset,
-                                       dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat img(dstSpecs.height, dstSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::merge(channels, img);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = dstSpecs.p1Offset + i * dstSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t dstPos = lineStart + j * 3;
-                    uint32_t p1Pos = srcSpecs.p3Offset + i * srcSpecs.p3Stride + j;
-                    uint32_t p2Pos = srcSpecs.p2Offset + i * srcSpecs.p2Stride + j;
-                    uint32_t p3Pos = srcSpecs.p1Offset + i * srcSpecs.p1Stride + j;
-                    outputFrame->getOffsetData()[dstPos + 0] = src[p1Pos];
-                    outputFrame->getOffsetData()[dstPos + 1] = src[p2Pos];
-                    outputFrame->getOffsetData()[dstPos + 2] = src[p3Pos];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p1Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p1Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p3Offset,
-                                       srcSpecs.p3Stride,
-                                       outputFrame->data() + dstSpecs.p1Offset,
-                                       dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            cv::Mat img(dstSpecs.height, dstSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::merge(channels, img);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStart = dstSpecs.p1Offset + i * dstSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t dstPos = lineStart + j * 3;
-                    uint32_t p1Pos = srcSpecs.p1Offset + i * srcSpecs.p1Stride + j;
-                    uint32_t p2Pos = srcSpecs.p2Offset + i * srcSpecs.p2Stride + j;
-                    uint32_t p3Pos = srcSpecs.p3Offset + i * srcSpecs.p3Stride + j;
-                    outputFrame->getOffsetData()[dstPos + 0] = src[p1Pos];
-                    outputFrame->getOffsetData()[dstPos + 1] = src[p2Pos];
-                    outputFrame->getOffsetData()[dstPos + 2] = src[p3Pos];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::RGB888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToBGR888u8(
-                src + srcSpecs.p1Offset, srcSpecs.width, srcSpecs.height, srcSpecs.p1Stride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat img(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat imgBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(img, imgBGR, cv::COLOR_RGB2BGR);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                uint32_t lineStartSrc = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                uint32_t lineStartDst = dstSpecs.p1Offset + i * dstSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    uint32_t dstPos = lineStartDst + j * 3;
-                    uint32_t srcPos = lineStartSrc + j * 3;
-                    outputFrame->getOffsetData()[dstPos + 0] = src[srcPos + 2];
-                    outputFrame->getOffsetData()[dstPos + 1] = src[srcPos + 1];
-                    outputFrame->getOffsetData()[dstPos + 2] = src[srcPos + 0];
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888i:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case dai::ImgFrame::Type::NV12: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PseudoPlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                                   src + srcSpecs.p2Offset,
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   srcSpecs.p1Stride,
-                                                   srcSpecs.p2Stride,
-                                                   ccAuxFrame->data(),
-                                                   auxStride);
-            fcvColorRGB888ToBGR888u8(
-                ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat frameUV(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC2, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            cv::Mat img(dstSpecs.height, dstSpecs.width, CV_8UC3, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColorTwoPlane(frameY, frameUV, img, cv::COLOR_YUV2BGR_NV12);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::YUV420p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                             src + srcSpecs.p2Offset,
-                                             src + srcSpecs.p3Offset,
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             srcSpecs.p1Stride,
-                                             srcSpecs.p2Stride,
-                                             srcSpecs.p3Stride,
-                                             ccAuxFrame->data(),
-                                             auxStride);
-            fcvColorRGB888ToBGR888u8(
-                ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartY = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartU = srcSpecs.p2Offset + (i / 2) * srcSpecs.p2Stride;
-                const uint32_t lineStartV = srcSpecs.p3Offset + (i / 2) * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + 3 * j;
-                    float Y = src[lineStartY + j];
-                    float U = src[lineStartU + (uint32_t)(j / 2)];
-                    float V = src[lineStartV + (uint32_t)(j / 2)];
-                    float R, G, B;
-                    RGBfromYUV(R, G, B, Y, U, V);
-                    outputFrame->getOffsetData()[pos + 0] = static_cast<uint8_t>(clampi(roundf(B), 0, 255.0f));
-                    outputFrame->getOffsetData()[pos + 1] = static_cast<uint8_t>(clampi(roundf(G), 0, 255.0f));
-                    outputFrame->getOffsetData()[pos + 2] = static_cast<uint8_t>(clampi(roundf(R), 0, 255.0f));
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAY8:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToNV12(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                       std::shared_ptr<ImageManipData> outputFrame,
-                                                                       FrameSpecs srcSpecs,
-                                                                       FrameSpecs dstSpecs,
-                                                                       dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::NV12;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-#endif
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p3Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p3Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p1Offset,
-                                       srcSpecs.p1Stride,
-                                       ccAuxFrame->data(),
-                                       auxStride);
-            fcvColorRGB888ToYCbCr420PseudoPlanaru8(ccAuxFrame->data(),
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   auxStride,
-                                                   outputFrame->data() + dstSpecs.p1Offset,
-                                                   outputFrame->data() + dstSpecs.p2Offset,
-                                                   dstSpecs.p1Stride,
-                                                   dstSpecs.p2Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartR = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartG = srcSpecs.p2Offset + i * srcSpecs.p2Stride;
-                const uint32_t lineStartB = srcSpecs.p3Offset + i * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2;
-                    const uint32_t p3Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2 + 1;
-                    float R = src[lineStartR + j];
-                    float G = src[lineStartG + j];
-                    float B = src[lineStartB + j];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p1Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p1Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p3Offset,
-                                       srcSpecs.p3Stride,
-                                       ccAuxFrame->data(),
-                                       auxStride);
-            fcvColorRGB888ToYCbCr420PseudoPlanaru8(ccAuxFrame->data(),
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   auxStride,
-                                                   outputFrame->data() + dstSpecs.p1Offset,
-                                                   outputFrame->data() + dstSpecs.p2Offset,
-                                                   dstSpecs.p1Stride,
-                                                   dstSpecs.p2Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartB = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartG = srcSpecs.p2Offset + i * srcSpecs.p2Stride;
-                const uint32_t lineStartR = srcSpecs.p3Offset + i * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2;
-                    const uint32_t p3Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2 + 1;
-                    float R = src[lineStartR + j];
-                    float G = src[lineStartG + j];
-                    float B = src[lineStartB + j];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::RGB888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToBGR888u8(src + srcSpecs.p1Offset, srcSpecs.width, srcSpecs.height, srcSpecs.p1Stride, ccAuxFrame->data(), auxStride);
-            fcvColorRGB888ToYCbCr420PseudoPlanaru8(ccAuxFrame->data(),
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   auxStride,
-                                                   outputFrame->data() + dstSpecs.p1Offset,
-                                                   outputFrame->data() + dstSpecs.p2Offset,
-                                                   dstSpecs.p1Stride,
-                                                   dstSpecs.p2Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = lineStart + j * 3;
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2;
-                    const uint32_t p3Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2 + 1;
-                    float R = src[pos + 0];
-                    float G = src[pos + 1];
-                    float B = src[pos + 2];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888i:
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToYCbCr420PseudoPlanaru8(src + srcSpecs.p1Offset,
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   srcSpecs.p1Stride,
-                                                   outputFrame->data() + dstSpecs.p1Offset,
-                                                   outputFrame->data() + dstSpecs.p2Offset,
-                                                   dstSpecs.p1Stride,
-                                                   dstSpecs.p2Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = lineStart + j * 3;
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2;
-                    const uint32_t p3Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2) * 2 + 1;
-                    float B = src[pos + 0];
-                    float G = src[pos + 1];
-                    float R = src[pos + 2];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        case dai::ImgFrame::Type::NV12:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case dai::ImgFrame::Type::YUV420p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                src + srcSpecs.p2Offset,
-                                srcSpecs.p2Stride,
-                                src + srcSpecs.p3Offset,
-                                srcSpecs.p3Stride,
-                                FASTCV_CHANNEL_0,
-                                FASTCV_IYUV,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelCombine2Planesu8(src + srcSpecs.p2Offset,
-                                       srcSpecs.width / 2,
-                                       srcSpecs.height / 2,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p3Offset,
-                                       srcSpecs.p3Stride,
-                                       outputFrame->data() + dstSpecs.p2Offset,
-                                       dstSpecs.p2Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(2);
-            channels.emplace_back(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            cv::Mat frameUV(dstSpecs.height / 2, dstSpecs.width / 2, CV_8UC2, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            cv::merge(channels, frameUV);
-            cv::Mat srcY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat dstY(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            srcY.copyTo(dstY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::GRAY8:
-            std::copy(src, src + inputSize, outputFrame->data());
-            memset(outputFrame->data() + dstSpecs.p2Offset, 128, dstSpecs.p2Stride * dstSpecs.height / 2);
-            done = true;
-            break;
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToYUV420p(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                          std::shared_ptr<ImageManipData> outputFrame,
-                                                                          FrameSpecs srcSpecs,
-                                                                          FrameSpecs dstSpecs,
-                                                                          dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::YUV420p;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-#endif
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p3Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p3Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p1Offset,
-                                       srcSpecs.p1Stride,
-                                       ccAuxFrame->data(),
-                                       auxStride);
-            fcvColorRGB888ToYCbCr420Planaru8(ccAuxFrame->data(),
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             auxStride,
-                                             outputFrame->data() + dstSpecs.p1Offset,
-                                             outputFrame->data() + dstSpecs.p2Offset,
-                                             outputFrame->data() + dstSpecs.p3Offset,
-                                             dstSpecs.p1Stride,
-                                             dstSpecs.p2Stride,
-                                             dstSpecs.p3Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartR = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartG = srcSpecs.p2Offset + i * srcSpecs.p2Stride;
-                const uint32_t lineStartB = srcSpecs.p3Offset + i * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2);
-                    const uint32_t p3Pos = dstSpecs.p3Offset + (i / 2) * dstSpecs.p3Stride + (j / 2);
-                    float R = src[lineStartR + j];
-                    float G = src[lineStartG + j];
-                    float B = src[lineStartB + j];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p1Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p1Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p3Offset,
-                                       srcSpecs.p3Stride,
-                                       ccAuxFrame->data(),
-                                       auxStride);
-            fcvColorRGB888ToYCbCr420Planaru8(ccAuxFrame->data(),
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             auxStride,
-                                             outputFrame->data() + dstSpecs.p1Offset,
-                                             outputFrame->data() + dstSpecs.p2Offset,
-                                             outputFrame->data() + dstSpecs.p3Offset,
-                                             dstSpecs.p1Stride,
-                                             dstSpecs.p2Stride,
-                                             dstSpecs.p3Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartR = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartG = srcSpecs.p2Offset + i * srcSpecs.p2Stride;
-                const uint32_t lineStartB = srcSpecs.p3Offset + i * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2);
-                    const uint32_t p3Pos = dstSpecs.p3Offset + (i / 2) * dstSpecs.p3Stride + (j / 2);
-                    float B = src[lineStartR + j];
-                    float G = src[lineStartG + j];
-                    float R = src[lineStartB + j];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::RGB888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToBGR888u8(src + srcSpecs.p1Offset, srcSpecs.width, srcSpecs.height, srcSpecs.p1Stride, ccAuxFrame->data(), auxStride);
-            fcvColorRGB888ToYCbCr420Planaru8(ccAuxFrame->data(),
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             auxStride,
-                                             outputFrame->data() + dstSpecs.p1Offset,
-                                             outputFrame->data() + dstSpecs.p2Offset,
-                                             outputFrame->data() + dstSpecs.p3Offset,
-                                             dstSpecs.p1Stride,
-                                             dstSpecs.p2Stride,
-                                             dstSpecs.p3Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = lineStart + j * 3;
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2);
-                    const uint32_t p3Pos = dstSpecs.p3Offset + (i / 2) * dstSpecs.p3Stride + (j / 2);
-                    float R = src[pos + 0];
-                    float G = src[pos + 1];
-                    float B = src[pos + 2];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888i:
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToYCbCr420Planaru8(src + srcSpecs.p1Offset,
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             srcSpecs.p1Stride,
-                                             outputFrame->data() + dstSpecs.p1Offset,
-                                             outputFrame->data() + dstSpecs.p2Offset,
-                                             outputFrame->data() + dstSpecs.p3Offset,
-                                             dstSpecs.p1Stride,
-                                             dstSpecs.p2Stride,
-                                             dstSpecs.p3Stride);
-#else
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStart = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = lineStart + j * 3;
-                    const uint32_t p1Pos = dstSpecs.p1Offset + i * dstSpecs.p1Stride + j;
-                    const uint32_t p2Pos = dstSpecs.p2Offset + (i / 2) * dstSpecs.p2Stride + (j / 2);
-                    const uint32_t p3Pos = dstSpecs.p3Offset + (i / 2) * dstSpecs.p3Stride + (j / 2);
-                    float B = src[pos + 0];
-                    float G = src[pos + 1];
-                    float R = src[pos + 2];
-                    float Y, U, V;
-                    YUVfromRGB(Y, U, V, R, G, B);
-                    outputFrame->getOffsetData()[p1Pos] = static_cast<uint8_t>(Y);
-                    if(i % 2 == 0 && j % 2 == 0) {
-                        outputFrame->getOffsetData()[p2Pos] = static_cast<uint8_t>(U);
-                        outputFrame->getOffsetData()[p3Pos] = static_cast<uint8_t>(V);
-                    }
-                }
-            }
-#endif
-            done = true;
-            break;
-        case dai::ImgFrame::Type::NV12: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                src + srcSpecs.p2Offset,
-                                srcSpecs.p2Stride,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_Y,
-                                FASTCV_NV12,
-                                outputFrame->data() + dstSpecs.p1Offset,
-                                dstSpecs.p1Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                src + srcSpecs.p2Offset,
-                                srcSpecs.p2Stride,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_U,
-                                FASTCV_NV12,
-                                outputFrame->data() + dstSpecs.p2Offset,
-                                dstSpecs.p2Stride);
-            fcvChannelExtractu8(src + srcSpecs.p1Offset,
-                                srcSpecs.width,
-                                srcSpecs.height,
-                                srcSpecs.p1Stride,
-                                src + srcSpecs.p2Offset,
-                                srcSpecs.p2Stride,
-                                0,
-                                0,
-                                FASTCV_CHANNEL_V,
-                                FASTCV_NV12,
-                                outputFrame->data() + dstSpecs.p3Offset,
-                                dstSpecs.p3Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameUV(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC2, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            std::vector<cv::Mat> channels;
-            channels.reserve(2);
-            channels.emplace_back(dstSpecs.height / 2, dstSpecs.width / 2, CV_8UC1, outputFrame->data() + dstSpecs.p2Offset, dstSpecs.p2Stride);
-            channels.emplace_back(dstSpecs.height / 2, dstSpecs.width / 2, CV_8UC1, outputFrame->data() + dstSpecs.p3Offset, dstSpecs.p3Stride);
-            cv::split(frameUV, channels);
-            cv::Mat srcY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat dstY(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            srcY.copyTo(dstY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::YUV420p:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAY8:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-bool ColorChange<ImageManipBuffer, ImageManipData>::colorConvertToGRAY8(const std::shared_ptr<ImageManipData> inputFrame,
-                                                                        std::shared_ptr<ImageManipData> outputFrame,
-                                                                        FrameSpecs srcSpecs,
-                                                                        FrameSpecs dstSpecs,
-                                                                        dai::ImgFrame::Type from) {
-    // dai::ImgFrame::Type to = dai::ImgFrame::Type::GRAY8;
-
-    auto src = inputFrame->getOffsetData().data();
-    auto inputSize = inputFrame->getOffsetSize();
-    uint32_t auxStride = ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-
-    bool done = false;
-    switch(from) {
-        case dai::ImgFrame::Type::RGB888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p1Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p1Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p3Offset,
-                                       srcSpecs.p3Stride,
-                                       ccAuxFrame->data(),
-                                       auxStride);
-            fcvColorRGB888ToGrayu8(ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            cv::Mat auxRGB(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::merge(channels, auxRGB);
-            // Convert to grayscale
-            cv::Mat gray(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(auxRGB, gray, cv::COLOR_RGB2GRAY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvChannelCombine3Planesu8(src + srcSpecs.p3Offset,
-                                       srcSpecs.width,
-                                       srcSpecs.height,
-                                       srcSpecs.p3Stride,
-                                       src + srcSpecs.p2Offset,
-                                       srcSpecs.p2Stride,
-                                       src + srcSpecs.p1Offset,
-                                       srcSpecs.p1Stride,
-                                       ccAuxFrame->data(),
-                                       auxStride);
-            fcvColorRGB888ToGrayu8(ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            std::vector<cv::Mat> channels;
-            channels.reserve(3);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            channels.emplace_back(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p3Offset), srcSpecs.p3Stride);
-            cv::Mat auxRGB(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::merge(channels, auxRGB);
-            cv::Mat gray(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(auxRGB, gray, cv::COLOR_BGR2GRAY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::RGB888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToGrayu8(
-                src + srcSpecs.p1Offset, srcSpecs.width, srcSpecs.height, srcSpecs.p1Stride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameRGB(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat gray(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(frameRGB, gray, cv::COLOR_RGB2GRAY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::BGR888i: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorRGB888ToBGR888u8(src + srcSpecs.p1Offset, srcSpecs.width, srcSpecs.height, srcSpecs.p1Stride, ccAuxFrame->data(), auxStride);
-            fcvColorRGB888ToGrayu8(ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat gray(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(frameBGR, gray, cv::COLOR_BGR2GRAY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::NV12: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PseudoPlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                                   src + srcSpecs.p2Offset,
-                                                   srcSpecs.width,
-                                                   srcSpecs.height,
-                                                   srcSpecs.p1Stride,
-                                                   srcSpecs.p2Stride,
-                                                   ccAuxFrame->data(),
-                                                   auxStride);
-            fcvColorRGB888ToGrayu8(ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            cv::Mat frameY(srcSpecs.height, srcSpecs.width, CV_8UC1, const_cast<uint8_t*>(src + srcSpecs.p1Offset), srcSpecs.p1Stride);
-            cv::Mat frameUV(srcSpecs.height / 2, srcSpecs.width / 2, CV_8UC2, const_cast<uint8_t*>(src + srcSpecs.p2Offset), srcSpecs.p2Stride);
-            cv::Mat auxBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::cvtColorTwoPlane(frameY, frameUV, auxBGR, cv::COLOR_YUV2BGR_NV12);
-            cv::Mat gray(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(auxBGR, gray, cv::COLOR_BGR2GRAY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::YUV420p: {
-#if defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-            fcvColorYCbCr420PlanarToRGB888u8(src + srcSpecs.p1Offset,
-                                             src + srcSpecs.p2Offset,
-                                             src + srcSpecs.p3Offset,
-                                             srcSpecs.width,
-                                             srcSpecs.height,
-                                             srcSpecs.p1Stride,
-                                             srcSpecs.p2Stride,
-                                             srcSpecs.p3Stride,
-                                             ccAuxFrame->data(),
-                                             auxStride);
-            fcvColorRGB888ToGrayu8(ccAuxFrame->data(), srcSpecs.width, srcSpecs.height, auxStride, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-#elif defined(DEPTHAI_HAVE_OPENCV_SUPPORT)
-            for(uint32_t i = 0; i < srcSpecs.height; ++i) {
-                const uint32_t lineStartY = srcSpecs.p1Offset + i * srcSpecs.p1Stride;
-                const uint32_t lineStartU = srcSpecs.p2Offset + (i / 2) * srcSpecs.p2Stride;
-                const uint32_t lineStartV = srcSpecs.p3Offset + (i / 2) * srcSpecs.p3Stride;
-                for(uint32_t j = 0; j < srcSpecs.width; ++j) {
-                    const uint32_t pos = srcSpecs.p1Offset + i * auxStride + 3 * j;
-                    float Y = src[lineStartY + j];
-                    float U = src[lineStartU + (uint32_t)(j / 2)];
-                    float V = src[lineStartV + (uint32_t)(j / 2)];
-                    float R, G, B;
-                    RGBfromYUV(R, G, B, Y, U, V);
-                    ccAuxFrame->data()[pos + 0] = static_cast<uint8_t>(clampi(roundf(B), 0, 255.0f));
-                    ccAuxFrame->data()[pos + 1] = static_cast<uint8_t>(clampi(roundf(G), 0, 255.0f));
-                    ccAuxFrame->data()[pos + 2] = static_cast<uint8_t>(clampi(roundf(R), 0, 255.0f));
-                }
-            }
-            cv::Mat auxBGR(srcSpecs.height, srcSpecs.width, CV_8UC3, ccAuxFrame->data(), auxStride);
-            cv::Mat gray(dstSpecs.height, dstSpecs.width, CV_8UC1, outputFrame->data() + dstSpecs.p1Offset, dstSpecs.p1Stride);
-            cv::cvtColor(auxBGR, gray, cv::COLOR_BGR2GRAY);
-#else
-            throw std::runtime_error("FastCV or OpenCV support required for this conversion");
-#endif
-            done = true;
-            break;
-        }
-        case dai::ImgFrame::Type::RAW8:
-        case dai::ImgFrame::Type::GRAY8:
-            std::copy(src, src + inputSize, outputFrame->data());
-            done = true;
-            break;
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-
-    return done;
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-void ColorChange<ImageManipBuffer, ImageManipData>::build(const FrameSpecs srcFrameSpecs,
-                                                          const FrameSpecs dstFrameSpecs,
-                                                          const ImgFrame::Type typeFrom,
-                                                          const ImgFrame::Type typeTo) {
-    from = typeFrom;
-    to = typeTo;
-    srcSpecs = srcFrameSpecs;
-    dstSpecs = dstFrameSpecs;
-    size_t newAuxFrameSize = ALIGN_UP(srcSpecs.height, DEPTHAI_HEIGHT_ALIGNMENT) * ALIGN_UP(3 * srcSpecs.width, DEPTHAI_STRIDE_ALIGNMENT);
-    if(!ccAuxFrame || ccAuxFrame->size() < newAuxFrameSize) ccAuxFrame = std::make_shared<ImageManipData>(newAuxFrameSize);
-}
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData>
-void ColorChange<ImageManipBuffer, ImageManipData>::apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst) {
-    float bpp;
-    int numPlanes;
-    getFrameTypeInfo(to, numPlanes, bpp);
-
-    bool done = false;
-    auto start = std::chrono::steady_clock::now();
-    switch(to) {
-        case dai::ImgFrame::Type::RGB888p:
-            done = colorConvertToRGB888p(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case dai::ImgFrame::Type::BGR888p:
-            done = colorConvertToBGR888p(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case dai::ImgFrame::Type::RGB888i:
-            done = colorConvertToRGB888i(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case dai::ImgFrame::Type::BGR888i:
-            done = colorConvertToBGR888i(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case dai::ImgFrame::Type::NV12:
-            done = colorConvertToNV12(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case dai::ImgFrame::Type::YUV420p:
-            done = colorConvertToYUV420p(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case dai::ImgFrame::Type::GRAY8:
-        case dai::ImgFrame::Type::RAW8:
-            done = colorConvertToGRAY8(src, dst, srcSpecs, dstSpecs, from);
-            break;
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-    auto diff = std::chrono::steady_clock::now() - start;
-    if(logger) logger->trace("ImageManip | colorConvert took {}ns", std::chrono::duration_cast<std::chrono::nanoseconds>(diff).count());
-
-    if(!done) {
-        if(logger) logger->error("Convert color from {} to {} not supported or failed.", (int)from, (int)to);
-        std::copy(src->getOffsetData().data(),
-                  src->getOffsetData().data() + (src->getOffsetSize() <= dst->getOffsetSize() ? src->getOffsetSize() : dst->getOffsetSize()),
-                  dst->getOffsetData().data());
-    }
-}
-
-//------------
-//--- Warp ---
-//------------
-
-// helper type for the visitor #4
-template <class... Ts>
-struct overloaded : Ts... {
-    using Ts::operator()...;
-};
-// explicit deduction guide (not needed as of C++20)
-template <class... Ts>
-overloaded(Ts...) -> overloaded<Ts...>;
-
-inline bool floatEq(float a, float b) {
-    return fabsf(a - b) <= 1e-6f;
-}
-
 inline bool isSingleChannelu8(const std::shared_ptr<dai::ImgFrame> img) {
     return img->getType() == dai::ImgFrame::Type::GRAY8 || img->getType() == dai::ImgFrame::Type::RAW8;
 }
 inline bool isSingleChannelu8(const dai::ImgFrame::Type type) {
     return type == dai::ImgFrame::Type::GRAY8 || type == dai::ImgFrame::Type::RAW8;
-}
-inline bool isSingleChannel(const dai::ImgFrame::Type type) {
-    return type == dai::ImgFrame::Type::GRAY8 || type == dai::ImgFrame::Type::RAW8 || type == dai::ImgFrame::Type::RAW16 || type == dai::ImgFrame::Type::GRAYF16
-           || type == dai::ImgFrame::Type::RAW32;
 }
 
 template <typename T>
@@ -2570,21 +632,6 @@ inline std::array<float, 2> matvecmul(std::array<std::array<float, 2>, 2> M, std
     return {x, y};
 }
 
-std::tuple<float, float, float, float> getOuterRect(const std::vector<std::array<float, 2>> points);
-
-std::vector<std::array<float, 2>> getHull(const std::vector<std::array<float, 2>> points);
-
-dai::RotatedRect getOuterRotatedRect(const std::vector<std::array<float, 2>>& points);
-
-std::array<std::array<float, 3>, 3> getResizeMat(Resize o, float width, float height, uint32_t outputWidth, uint32_t outputHeight);
-
-void getTransformImpl(const ManipOp& op,
-                      std::array<std::array<float, 3>, 3>& transform,
-                      std::array<std::array<float, 2>, 4>& imageCorners,
-                      std::vector<std::array<std::array<float, 2>, 4>>& srcCorners,
-                      uint32_t& outputWidth,
-                      uint32_t& outputHeight);
-
 template <typename C>
 std::tuple<std::array<std::array<float, 3>, 3>, std::array<std::array<float, 2>, 4>, std::vector<std::array<std::array<float, 2>, 4>>> getTransform(
     const C& ops, uint32_t inputWidth, uint32_t inputHeight, uint32_t outputWidth, uint32_t outputHeight) {
@@ -2596,14 +643,6 @@ std::tuple<std::array<std::array<float, 3>, 3>, std::array<std::array<float, 2>,
     }
     return {transform, imageCorners, srcCorners};
 }
-
-void getOutputSizeFromCorners(const std::array<std::array<float, 2>, 4>& corners,
-                              const bool center,
-                              const std::array<std::array<float, 3>, 3> transformInv,
-                              const uint32_t srcWidth,
-                              const uint32_t srcHeight,
-                              uint32_t& outputWidth,
-                              uint32_t& outputHeight);
 
 template <typename C>
 std::tuple<std::array<std::array<float, 3>, 3>, std::array<std::array<float, 2>, 4>, std::vector<std::array<std::array<float, 2>, 4>>> getFullTransform(
@@ -2675,8 +714,8 @@ inline dai::ImgFrame::Type getValidType(dai::ImgFrame::Type type) {
     return isSingleChannelu8(type) ? VALID_TYPE_GRAY : VALID_TYPE_COLOR;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>& ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::build(
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>& ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::build(
     const ImageManipOpsBase<Container>& newBase, ImgFrame::Type outType, FrameSpecs srcFrameSpecs, ImgFrame::Type inFrameType) {
     const auto newCfgStr = newBase.str();
     if(outType == ImgFrame::Type::NONE) {
@@ -2781,8 +820,8 @@ ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>& ImageManipO
     return *this;
 }  // namespace impl
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>& ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::buildUndistort(
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>& ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::buildUndistort(
     bool enable,
     const std::array<float, 9>& cameraMatrix,
     const std::array<float, 9>& newCameraMatrix,
@@ -2796,10 +835,8 @@ ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>& ImageManipO
     return *this;
 }
 
-size_t getFrameSize(const ImgFrame::Type type, const FrameSpecs& specs);
-
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-bool ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst) {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+bool ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst) {
     size_t requiredSize = getFrameSize(inType, srcSpecs);
     if(src->getOffsetSize() < requiredSize)
         throw std::runtime_error("ImageManip not built for the source image specs. Consider rebuilding with the new configuration.");
@@ -2816,9 +853,9 @@ bool ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::apply(
                              base.colormap != Colormap::NONE ? colormapFrame : (type == outputFrameType ? dst : warpedFrame));
         }
         if(mode & MODE_COLORMAP) {
-            uint8_t* colormapDst = outputFrameType == VALID_TYPE_COLOR ? dst->getOffsetData().data() : warpedFrame->data();
+            uint8_t* colormapDst = outputFrameType == VALID_TYPE_COLOR ? dst->getOffsetData().data() : warpedFrame->getOffsetData().data();
             size_t colormapDstStride = outputFrameType == VALID_TYPE_COLOR ? getOutputStride() : ALIGN_UP(base.outputWidth, DEPTHAI_STRIDE_ALIGNMENT);
-            uint8_t* colormapSrc = mode & MODE_WARP ? colormapFrame->data() : (convertInput ? convertedFrame->data() : src->getOffsetData().data());
+            uint8_t* colormapSrc = mode & MODE_WARP ? colormapFrame->getOffsetData().data() : (convertInput ? convertedFrame->getOffsetData().data() : src->getOffsetData().data());
             size_t colormapSrcStride = !(mode & MODE_WARP) && !convertInput ? srcSpecs.p1Stride : ALIGN_UP(base.outputWidth, DEPTHAI_STRIDE_ALIGNMENT);
             cv::Mat gray(base.outputWidth, base.outputHeight, CV_8UC1, colormapSrc, colormapSrcStride);
             cv::Mat color(base.outputWidth, base.outputHeight, CV_8UC3, colormapDst, colormapDstStride);
@@ -2848,18 +885,18 @@ bool ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::apply(
 #endif
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getOutputWidth() const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+size_t ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getOutputWidth() const {
     return base.outputWidth;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getOutputHeight() const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+size_t ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getOutputHeight() const {
     return base.outputHeight;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getOutputStride(uint8_t plane) const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+size_t ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getOutputStride(uint8_t plane) const {
     if(mode == 0) return plane == 0 ? srcSpecs.p1Stride : (plane == 1 ? srcSpecs.p2Stride : (plane == 2 ? srcSpecs.p3Stride : 0));
     auto specs = getOutputFrameSpecs(outputFrameType);
     if(plane == 0)
@@ -2872,8 +909,8 @@ size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getO
         return 0;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getOutputPlaneSize(uint8_t plane) const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+size_t ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getOutputPlaneSize(uint8_t plane) const {
     if(mode == 0) return 0;
     size_t size = 0;
     switch(outputFrameType) {
@@ -2935,8 +972,8 @@ size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getO
     return size;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getOutputSize() const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+size_t ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getOutputSize() const {
     if(mode == 0) return 0;
     size_t size = 0;
     switch(outputFrameType) {
@@ -2988,16 +1025,16 @@ size_t ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getO
     return size;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-FrameSpecs ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getOutputFrameSpecs(ImgFrame::Type type) const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+FrameSpecs ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getOutputFrameSpecs(ImgFrame::Type type) const {
     if(mode == 0)
         return srcSpecs;
     else
         return getDstFrameSpecs(base.outputWidth, base.outputHeight, type);
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-std::vector<RotatedRect> ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getSrcCrops() const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+std::vector<RotatedRect> ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getSrcCrops() const {
     std::vector<RotatedRect> crops;
     for(const auto& corners : srcCorners) {
         auto rect = getOuterRotatedRect({corners[0], corners[1], corners[2], corners[3]});
@@ -3006,13 +1043,13 @@ std::vector<RotatedRect> ImageManipOperations<ImageManipBuffer, ImageManipData, 
     return crops;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-std::array<std::array<float, 3>, 3> ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::getMatrix() const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+std::array<std::array<float, 3>, 3> ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::getMatrix() const {
     return matrix;
 }
 
-template <template <typename T> typename ImageManipBuffer, typename ImageManipData, typename WarpBackend>
-std::string ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>::toString() const {
+template <typename ImageManipData, typename ColorChangeBackend, typename WarpBackend>
+std::string ImageManipOperations<ImageManipData, ColorChangeBackend, WarpBackend>::toString() const {
     std::stringstream cStr;
     cStr << getConfigString(base);
     if(outputOps.size() > 0) {
@@ -3023,575 +1060,6 @@ std::string ImageManipOperations<ImageManipBuffer, ImageManipData, WarpBackend>:
         }
     }
     return cStr.str();
-}
-
-void WarpH::build(const FrameSpecs srcFrameSpecs,
-                  const FrameSpecs dstFrameSpecs,
-                  const ImgFrame::Type type,
-                  const std::array<std::array<float, 3>, 3> matrix,
-                  std::vector<std::array<std::array<float, 2>, 4>> srcCorners) {
-    this->matrix = matrix;
-    this->type = type;
-    this->srcSpecs = srcFrameSpecs;
-    this->dstSpecs = dstFrameSpecs;
-
-    if(!fastCvBorder || fastCvBorder->size() < this->dstSpecs.height * 2)
-        fastCvBorder = std::make_shared<_ImageManipBuffer<uint32_t>>(this->dstSpecs.height * 2);
-
-    const uint32_t inWidth = srcFrameSpecs.width;
-    const uint32_t inHeight = srcFrameSpecs.height;
-    this->sourceMinX = 0;
-    this->sourceMaxX = inWidth;
-    this->sourceMinY = 0;
-    this->sourceMaxY = inHeight;
-    for(const auto& corners : srcCorners) {
-        auto [minx, maxx, miny, maxy] = getOuterRect(std::vector<std::array<float, 2>>(corners.begin(), corners.end()));
-        this->sourceMinX = std::max(this->sourceMinX, (size_t)std::floor(std::max(minx, 0.f)));
-        this->sourceMinY = std::max(this->sourceMinY, (size_t)std::floor(std::max(miny, 0.f)));
-        this->sourceMaxX = std::min(this->sourceMaxX, (size_t)std::ceil(maxx));
-        this->sourceMaxY = std::min(this->sourceMaxY, (size_t)std::ceil(maxy));
-    }
-    if(this->sourceMinX >= this->sourceMaxX || this->sourceMinY >= this->sourceMaxY) throw std::runtime_error("Initial crop is outside the source image");
-}
-
-void WarpH::buildUndistort(bool enable,
-                           const std::array<float, 9>& cameraMatrix,
-                           const std::array<float, 9>& newCameraMatrix,
-                           const std::vector<float>& distCoeffs,
-                           const ImgFrame::Type type,
-                           const uint32_t srcWidth,
-                           const uint32_t srcHeight,
-                           const uint32_t dstWidth,
-                           const uint32_t dstHeight) {
-#ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-    if(enable) {
-        if(!undistortImpl) undistortImpl = std::make_unique<UndistortOpenCvImpl>(this->logger);
-        auto undistortStatus = undistortImpl->build(cameraMatrix, newCameraMatrix, distCoeffs, type, srcWidth, srcHeight, dstWidth, dstHeight);
-        switch(undistortStatus) {
-            case UndistortOpenCvImpl::BuildStatus::ONE_SHOT:
-                this->enableUndistort = true;
-                this->undistortOneShot = true;
-                break;
-            case UndistortOpenCvImpl::BuildStatus::TWO_SHOT:
-                this->enableUndistort = true;
-                this->undistortOneShot = false;
-                break;
-            case UndistortOpenCvImpl::BuildStatus::NOT_USED:
-                this->enableUndistort = false;
-                this->undistortOneShot = false;
-                break;
-            case UndistortOpenCvImpl::BuildStatus::NOT_BUILT:
-                break;
-            case UndistortOpenCvImpl::BuildStatus::ERROR:
-                this->enableUndistort = false;
-                break;
-        }
-
-        if(this->enableUndistort && !this->undistortOneShot) {
-            auto frameSize = getAlignedOutputFrameSize(type, srcWidth, srcHeight);
-            if(!auxFrame || auxFrame->getOffsetSize() < frameSize) {
-                // When undistort is needed but cannot one shot - undistorted frame must be written to an aux buffer
-                auxFrame = std::make_shared<_ImageManipMemory>(frameSize);
-            }
-        }
-    } else {
-        undistortImpl = nullptr;
-        this->enableUndistort = false;
-    }
-#else
-    (void)enable;
-    (void)cameraMatrix;
-    (void)newCameraMatrix;
-    (void)distCoeffs;
-    (void)type;
-    (void)srcWidth;
-    (void)srcHeight;
-    (void)dstWidth;
-    (void)dstHeight;
-    throw std::runtime_error("Undistort requires OpenCV support");
-#endif
-}
-
-void WarpH::transform(const std::shared_ptr<OffsetMemory>& src,
-                      std::shared_ptr<OffsetMemory> dst,
-                      const size_t srcWidth,
-                      const size_t srcHeight,
-                      const size_t srcStride,
-                      const size_t dstWidth,
-                      const size_t dstHeight,
-                      const size_t dstStride,
-                      const uint16_t numChannels,
-                      const uint16_t bpp,
-                      const std::array<std::array<float, 3>, 3> matrix,
-                      const std::vector<uint32_t>& background) {
-    if(1) {
-#ifdef DEPTHAI_IMAGEMANIPV2_OPENCV
-        transformOpenCV(src->getOffsetData().data(),
-                        dst->getOffsetData().data(),
-                        srcWidth,
-                        srcHeight,
-                        srcStride,
-                        dstWidth,
-                        dstHeight,
-                        dstStride,
-                        numChannels,
-                        bpp,
-                        matrix,
-                        background,
-                        this->srcSpecs,
-                        this->sourceMinX,
-                        this->sourceMinY,
-                        this->sourceMaxX,
-                        this->sourceMaxY);
-#else
-        throw std::runtime_error("OpenCV backend not available");
-#endif
-    } else {
-#ifdef DEPTHAI_IMAGEMANIPV2_FASTCV
-        transformFastCV(src->getOffsetData().data()),
-                        dst->getOffsetData().data(),
-                        srcWidth,
-                        srcHeight,
-                        srcStride,
-                        dstWidth,
-                        dstHeight,
-                        dstStride,
-                        numChannels,
-                        bpp,
-                        matrix,
-                        background,
-                        this->srcSpecs,
-                        this->sourceMinX,
-                        this->sourceMinY,
-                        this->sourceMaxX,
-                        this->sourceMaxY,
-                        fastCvBorder->data());
-#else
-        throw std::runtime_error("FastCV backend not available");
-#endif
-    }
-
-#if !defined(DEPTHAI_HAVE_OPENCV_SUPPORT) && !defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    (void)src;
-    (void)dst;
-    (void)srcWidth;
-    (void)srcHeight;
-    (void)srcStride;
-    (void)dstWidth;
-    (void)dstHeight;
-    (void)dstStride;
-    (void)numChannels;
-    (void)bpp;
-    (void)matrix;
-    (void)background;
-#endif
-}
-
-void printSpecs(spdlog::async_logger& logger, FrameSpecs specs);
-
-void WarpH::apply(const std::shared_ptr<OffsetMemory>& src, std::shared_ptr<OffsetMemory> dst) {
-    auto undistortDst = this->isIdentityWarp() || this->undistortOneShot ? dst : auxFrame;
-    auto undistortSpecs =
-        this->isIdentityWarp() || this->undistortOneShot ? this->dstSpecs : getDstFrameSpecs(this->srcSpecs.width, this->srcSpecs.height, this->type);
-    auto warpSrc = this->enableUndistort ? auxFrame : src;
-    auto warpSrcSpecs = this->enableUndistort ? undistortSpecs : this->srcSpecs;
-    // Apply transformation multiple times depending on the image format
-    switch(this->type) {
-        case ImgFrame::Type::RGB888i:
-        case ImgFrame::Type::BGR888i:
-#if DEPTHAI_IMAGEMANIPV2_OPENCV && defined(DEPTHAI_HAVE_OPENCV_SUPPORT) || DEPTHAI_IMAGEMANIPV2_FASTCV && defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    #ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-            if(this->enableUndistort) {
-                cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC3, src->offset(this->srcSpecs.p1Offset)->getOffsetData().data(), this->srcSpecs.p1Stride);
-                cv::Mat dstCv(
-                    undistortSpecs.height, undistortSpecs.width, CV_8UC3, undistortDst->offset(undistortSpecs.p1Offset)->getOffsetData().data(), undistortSpecs.p1Stride);
-                this->undistortImpl->undistort(srcCv, dstCv);
-            }
-    #endif
-            if(!this->undistortOneShot && !this->isIdentityWarp()) {
-                transform(warpSrc->offset(warpSrcSpecs.p1Offset),
-                          dst->offset(this->dstSpecs.p1Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p1Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p1Stride,
-                          3,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[0], this->backgroundColor[1], this->backgroundColor[2]});
-            }
-#else
-            (void)src;
-            (void)dst;
-            throw std::runtime_error("OpenCV or FastCV backend not available");
-#endif
-            break;
-        case ImgFrame::Type::BGR888p:
-        case ImgFrame::Type::RGB888p:
-#if DEPTHAI_IMAGEMANIPV2_OPENCV && defined(DEPTHAI_HAVE_OPENCV_SUPPORT) || DEPTHAI_IMAGEMANIPV2_FASTCV && defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    #ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-            if(this->enableUndistort) {
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC1, src->offset(this->srcSpecs.p1Offset)->getOffsetData().data(), this->srcSpecs.p1Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_8UC1, undistortDst->offset(undistortSpecs.p1Offset)->getOffsetData().data(), undistortSpecs.p1Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC1, src->offset(this->srcSpecs.p2Offset)->getOffsetData().data(), this->srcSpecs.p2Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_8UC1, undistortDst->offset(undistortSpecs.p2Offset)->getOffsetData().data(), undistortSpecs.p2Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC1, src->offset(this->srcSpecs.p3Offset)->getOffsetData().data(), this->srcSpecs.p3Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_8UC1, undistortDst->offset(undistortSpecs.p3Offset)->getOffsetData().data(), undistortSpecs.p3Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-            }
-    #endif
-            if(!this->undistortOneShot && !this->isIdentityWarp()) {
-                transform(warpSrc->offset(warpSrcSpecs.p1Offset),
-                          dst->offset(this->dstSpecs.p1Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p1Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p1Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[0]});
-                transform(warpSrc->offset(warpSrcSpecs.p2Offset),
-                          dst->offset(this->dstSpecs.p2Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p2Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p2Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[1]});
-                transform(warpSrc->offset(warpSrcSpecs.p3Offset),
-                          dst->offset(this->dstSpecs.p3Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p3Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p3Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[2]});
-            }
-#else
-            (void)src;
-            (void)dst;
-            throw std::runtime_error("OpenCV or FastCV backend not available");
-#endif
-            break;
-        case ImgFrame::Type::YUV420p:
-#if DEPTHAI_IMAGEMANIPV2_OPENCV && defined(DEPTHAI_HAVE_OPENCV_SUPPORT) || DEPTHAI_IMAGEMANIPV2_FASTCV && defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    #ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-            if(this->enableUndistort) {
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC1, src->offset(this->srcSpecs.p1Offset)->getOffsetData().data(), this->srcSpecs.p1Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_8UC1, undistortDst->offset(undistortSpecs.p1Offset)->getOffsetData().data(), undistortSpecs.p1Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-                {
-                    cv::Mat srcCv(
-                        this->srcSpecs.height / 2, this->srcSpecs.width / 2, CV_8UC1, src->offset(this->srcSpecs.p2Offset)->getOffsetData().data(), this->srcSpecs.p2Stride);
-                    cv::Mat dstCv(undistortSpecs.height / 2,
-                                  undistortSpecs.width / 2,
-                                  CV_8UC1,
-                                  undistortDst->offset(undistortSpecs.p2Offset)->getOffsetData().data(),
-                                  undistortSpecs.p2Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-                {
-                    cv::Mat srcCv(
-                        this->srcSpecs.height / 2, this->srcSpecs.width / 2, CV_8UC1, src->offset(this->srcSpecs.p3Offset)->getOffsetData().data(), this->srcSpecs.p3Stride);
-                    cv::Mat dstCv(undistortSpecs.height / 2,
-                                  undistortSpecs.width / 2,
-                                  CV_8UC1,
-                                  undistortDst->offset(undistortSpecs.p3Offset)->getOffsetData().data(),
-                                  undistortSpecs.p3Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-            }
-    #endif
-            if(!this->undistortOneShot && !this->isIdentityWarp()) {
-                transform(warpSrc->offset(warpSrcSpecs.p1Offset),
-                          dst->offset(this->dstSpecs.p1Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p1Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p1Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[0]});
-                transform(warpSrc->offset(warpSrcSpecs.p2Offset),
-                          dst->offset(this->dstSpecs.p2Offset),
-                          warpSrcSpecs.width / 2,
-                          warpSrcSpecs.height / 2,
-                          warpSrcSpecs.p2Stride,
-                          this->dstSpecs.width / 2,
-                          this->dstSpecs.height / 2,
-                          this->dstSpecs.p2Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[1]});
-                transform(warpSrc->offset(warpSrcSpecs.p3Offset),
-                          dst->offset(this->dstSpecs.p3Offset),
-                          warpSrcSpecs.width / 2,
-                          warpSrcSpecs.height / 2,
-                          warpSrcSpecs.p3Stride,
-                          this->dstSpecs.width / 2,
-                          this->dstSpecs.height / 2,
-                          this->dstSpecs.p3Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[2]});
-            }
-#else
-            (void)src;
-            (void)dst;
-            throw std::runtime_error("OpenCV or FastCV backend not available");
-#endif
-            break;
-        case ImgFrame::Type::NV12:
-#if DEPTHAI_IMAGEMANIPV2_OPENCV && defined(DEPTHAI_HAVE_OPENCV_SUPPORT) || DEPTHAI_IMAGEMANIPV2_FASTCV && defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    #ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-            if(this->enableUndistort) {
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC1, src->offset(this->srcSpecs.p1Offset)->getOffsetData().data(), this->srcSpecs.p1Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_8UC1, undistortDst->offset(undistortSpecs.p1Offset)->getOffsetData().data(), undistortSpecs.p1Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-                {
-                    cv::Mat srcCv(
-                        this->srcSpecs.height / 2, this->srcSpecs.width / 2, CV_8UC2, src->offset(this->srcSpecs.p2Offset)->getOffsetData().data(), this->srcSpecs.p2Stride);
-                    cv::Mat dstCv(undistortSpecs.height / 2,
-                                  undistortSpecs.width / 2,
-                                  CV_8UC2,
-                                  undistortDst->offset(undistortSpecs.p2Offset)->getOffsetData().data(),
-                                  undistortSpecs.p2Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-            }
-    #endif
-            if(!this->undistortOneShot && !this->isIdentityWarp()) {
-                transform(warpSrc->offset(warpSrcSpecs.p1Offset),
-                          dst->offset(this->dstSpecs.p1Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p1Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p1Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[0]});
-                transform(warpSrc->offset(warpSrcSpecs.p2Offset),
-                          dst->offset(this->dstSpecs.p2Offset),
-                          warpSrcSpecs.width / 2,
-                          warpSrcSpecs.height / 2,
-                          warpSrcSpecs.p2Stride,
-                          this->dstSpecs.width / 2,
-                          this->dstSpecs.height / 2,
-                          this->dstSpecs.p2Stride,
-                          2,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[1], this->backgroundColor[2]});
-            }
-#else
-            (void)src;
-            (void)dst;
-            throw std::runtime_error("OpenCV or FastCV backend not available");
-#endif
-            break;
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::GRAY8:
-#if DEPTHAI_IMAGEMANIPV2_OPENCV && defined(DEPTHAI_HAVE_OPENCV_SUPPORT) || DEPTHAI_IMAGEMANIPV2_FASTCV && defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    #ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-            if(this->enableUndistort) {
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_8UC1, src->offset(this->srcSpecs.p1Offset)->getOffsetData().data(), this->srcSpecs.p1Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_8UC1, undistortDst->offset(undistortSpecs.p1Offset)->getOffsetData().data(), undistortSpecs.p1Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-            }
-    #endif
-            if(!this->undistortOneShot && !this->isIdentityWarp()) {
-                transform(warpSrc->offset(warpSrcSpecs.p1Offset),
-                          dst->offset(this->dstSpecs.p1Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p1Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p1Stride,
-                          1,
-                          1,
-                          this->matrix,
-                          {this->backgroundColor[0]});
-            }
-#else
-            (void)src;
-            (void)dst;
-            throw std::runtime_error("OpenCV or FastCV backend not available");
-#endif
-            break;
-        case ImgFrame::Type::RAW16:
-#if DEPTHAI_IMAGEMANIPV2_OPENCV && defined(DEPTHAI_HAVE_OPENCV_SUPPORT) || DEPTHAI_IMAGEMANIPV2_FASTCV && defined(DEPTHAI_HAVE_FASTCV_SUPPORT)
-    #ifdef DEPTHAI_HAVE_OPENCV_SUPPORT
-            if(this->enableUndistort) {
-                {
-                    cv::Mat srcCv(this->srcSpecs.height, this->srcSpecs.width, CV_16UC1, src->offset(this->srcSpecs.p1Offset)->getOffsetData().data(), this->srcSpecs.p1Stride);
-                    cv::Mat dstCv(
-                        undistortSpecs.height, undistortSpecs.width, CV_16UC1, undistortDst->offset(undistortSpecs.p1Offset)->getOffsetData().data(), undistortSpecs.p1Stride);
-                    this->undistortImpl->undistort(srcCv, dstCv);
-                }
-            }
-    #endif
-            if(!this->undistortOneShot && !this->isIdentityWarp()) {
-                transform(warpSrc->offset(warpSrcSpecs.p1Offset),
-                          dst->offset(this->dstSpecs.p1Offset),
-                          warpSrcSpecs.width,
-                          warpSrcSpecs.height,
-                          warpSrcSpecs.p1Stride,
-                          this->dstSpecs.width,
-                          this->dstSpecs.height,
-                          this->dstSpecs.p1Stride,
-                          1,
-                          2,
-                          this->matrix,
-                          {this->backgroundColor[0]});
-            }
-#else
-            (void)src;
-            (void)dst;
-            throw std::runtime_error("OpenCV or FastCV backend not available");
-#endif
-            break;
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            throw std::runtime_error("Unsupported image format. Only YUV420p, RGB888p, BGR888p, RGB888i, BGR888i, RAW8, NV12, GRAY8 are supported");
-            break;
-    }
-
-#ifndef DEPTHAI_HAVE_OPENCV_SUPPORT
-    (void)warpSrcSpecs;
-#endif
-}
-
-bool Warp::isIdentityWarp() const {
-    return (matrix[0][0] == 1.0f && matrix[0][1] == 0.0f && matrix[0][2] == 0.0f && matrix[1][0] == 0.0f && matrix[1][1] == 1.0f && matrix[1][2] == 0.0f
-            && matrix[2][0] == 0.0f && matrix[2][1] == 0.0f && matrix[2][2] == 1.0f)
-           && (srcSpecs.width == dstSpecs.width && srcSpecs.height == dstSpecs.height);
-}
-
-Warp& Warp::setBackgroundColor(const uint32_t r, const uint32_t g, const uint32_t b) {
-    background = ImageManipOpsBase<Container>::Background::COLOR;
-    switch(type) {
-        case ImgFrame::Type::YUV420p:
-        case ImgFrame::Type::NV12: {
-            float y, u, v;
-            YUVfromRGB(y, u, v, r, g, b);
-            backgroundColor[0] = std::round(y);
-            backgroundColor[1] = std::round(u);
-            backgroundColor[2] = std::round(v);
-            break;
-        }
-        case ImgFrame::Type::RGB888p:
-        case ImgFrame::Type::RGB888i:
-            backgroundColor[0] = r;
-            backgroundColor[1] = g;
-            backgroundColor[2] = b;
-            break;
-        case ImgFrame::Type::BGR888p:
-        case ImgFrame::Type::BGR888i:
-            backgroundColor[0] = b;
-            backgroundColor[1] = g;
-            backgroundColor[2] = r;
-            break;
-        case ImgFrame::Type::RAW8:
-        case ImgFrame::Type::GRAY8:
-            // backgroundColor[0] = 0.299f * r + 0.587f * g + 0.114f * b;
-            backgroundColor[0] = b;
-            break;
-        case ImgFrame::Type::RAW16:
-            backgroundColor[0] = r;
-            break;
-        case ImgFrame::Type::YUV422i:
-        case ImgFrame::Type::YUV444p:
-        case ImgFrame::Type::YUV422p:
-        case ImgFrame::Type::YUV400p:
-        case ImgFrame::Type::RGBA8888:
-        case ImgFrame::Type::RGB161616:
-        case ImgFrame::Type::LUT2:
-        case ImgFrame::Type::LUT4:
-        case ImgFrame::Type::LUT16:
-        case ImgFrame::Type::RAW14:
-        case ImgFrame::Type::RAW12:
-        case ImgFrame::Type::RAW10:
-        case ImgFrame::Type::PACK10:
-        case ImgFrame::Type::PACK12:
-        case ImgFrame::Type::YUV444i:
-        case ImgFrame::Type::NV21:
-        case ImgFrame::Type::BITSTREAM:
-        case ImgFrame::Type::HDR:
-        case ImgFrame::Type::RGBF16F16F16p:
-        case ImgFrame::Type::BGRF16F16F16p:
-        case ImgFrame::Type::RGBF16F16F16i:
-        case ImgFrame::Type::BGRF16F16F16i:
-        case ImgFrame::Type::GRAYF16:
-        case ImgFrame::Type::RAW32:
-        case ImgFrame::Type::NONE:
-            break;
-    }
-    return *this;
 }
 
 }  // namespace impl
